+++ dash_core_components/Graph.py
@@ -27,20 +27,13 @@ plotly.js's `animate` function
 - hoverData (dict; optional): Data from latest hover event
 - config (optional): Plotly.js config options.
 See https://plot.ly/javascript/configuration-options/
-for more info.. config has the following type: dict containing keys 'topojsonURL', 'fillFrame', 'displayModeBar', 'autosizable', 'doubleClick', 'sendData', 'queueLength', 'displaylogo', 'edits', 'showLink', 'showAxisDragHandles', 'modeBarButtons', 'scrollZoom', 'editable', 'linkText', 'staticPlot', 'showTips', 'mapboxAccessToken', 'frameMargins', 'plotGlPixelRatio', 'modeBarButtonsToRemove', 'modeBarButtonsToAdd', 'showAxisRangeEntryBoxes'.
-Those keys have the following types: 
-  - topojsonURL (string; optional): URL to topojson files used in geo charts
-  - fillFrame (boolean; optional): if we DO autosize, do we fill the container or the screen?
-  - displayModeBar (a value equal to: true, false, 'hover'; optional): display the mode bar (true, false, or 'hover')
-  - autosizable (boolean; optional): DO autosize once regardless of layout.autosize
-(use default width or height values otherwise)
-  - doubleClick (a value equal to: false, 'reset', 'autosize', 'reset+autosize'; optional): double click interaction (false, 'reset', 'autosize' or 'reset+autosize')
-  - sendData (boolean; optional): if we show a link, does it contain data or just link to a plotly file?
-  - queueLength (number; optional): set the length of the undo/redo queue
-  - displaylogo (boolean; optional): add the plotly logo on the end of the mode bar
-  - edits (optional): a set of editable properties. edits has the following type: dict containing keys 'colorbarPosition', 'annotationPosition', 'shapePosition', 'annotationText', 'titleText', 'legendPosition', 'legendText', 'annotationTail', 'colorbarTitleText', 'axisTitleText'.
-Those keys have the following types: 
-  - colorbarPosition (boolean; optional)
+for more info.. config has the following type: dict containing keys 'staticPlot', 'editable', 'edits', 'autosizable', 'queueLength', 'fillFrame', 'frameMargins', 'scrollZoom', 'doubleClick', 'showTips', 'showAxisDragHandles', 'showAxisRangeEntryBoxes', 'showLink', 'sendData', 'linkText', 'displayModeBar', 'modeBarButtonsToRemove', 'modeBarButtonsToAdd', 'modeBarButtons', 'displaylogo', 'plotGlPixelRatio', 'topojsonURL', 'mapboxAccessToken'.
+Those keys have the following types:
+  - staticPlot (boolean; optional): no interactivity, for export or image generation
+  - editable (boolean; optional): we can edit titles, move annotations, etc - sets all pieces of `edits`
+unless a separate `edits` config item overrides individual parts
+  - edits (optional): a set of editable properties. edits has the following type: dict containing keys 'annotationPosition', 'annotationTail', 'annotationText', 'axisTitleText', 'colorbarPosition', 'colorbarTitleText', 'legendPosition', 'legendText', 'shapePosition', 'titleText'.
+Those keys have the following types:
   - annotationPosition (boolean; optional): annotationPosition: the main anchor of the annotation, which is the
 text (if no arrow) or the arrow (which drags the whole thing leaving
 the arrow length & direction unchanged)
@@ -78,22 +71,22 @@ Common names include:
 - (Geo): zoomInGeo, zoomOutGeo, resetGeo, hoverClosestGeo
 - hoverClosestGl2d, hoverClosestPie, toggleHover, resetViews
   - modeBarButtonsToAdd (list; optional): add mode bar button using config objects
-  - showAxisRangeEntryBoxes (boolean; optional): enable direct range entry at the pan/zoom drag points
-(drag handles must be enabled above)
-- id (string; optional): The ID of this component, used to identify dash components
-in callbacks. The ID needs to be unique across all of the
-components in an app.
-- selectedData (dict; optional): Data from latest select event
-
-Available events: 'click', 'clickannotation', 'hover', 'selected', 'relayout', 'unhover'"""
+  - modeBarButtons (boolean | number | string | dict | list; optional): fully custom mode bar buttons as nested array,
+where the outer arrays represents button groups, and
+the inner arrays have buttons config objects or names of default buttons
+  - displaylogo (boolean; optional): add the plotly logo on the end of the mode bar
+  - plotGlPixelRatio (number; optional): increase the pixel ratio for Gl plot images
+  - topojsonURL (string; optional): URL to topojson files used in geo charts
+  - mapboxAccessToken (boolean | number | string | dict | list; optional): Mapbox access token (required to plot mapbox trace types)
+If using an Mapbox Atlas server, set this option to '',
+so that plotly.js won't attempt to authenticate to the public Mapbox server."""
     @_explicitize_args
     def __init__(self, style=Component.UNDEFINED, figure=Component.UNDEFINED, animation_options=Component.UNDEFINED, className=Component.UNDEFINED, clear_on_unhover=Component.UNDEFINED, relayoutData=Component.UNDEFINED, config=Component.UNDEFINED, clickData=Component.UNDEFINED, animate=Component.UNDEFINED, hoverData=Component.UNDEFINED, clickAnnotationData=Component.UNDEFINED, id=Component.UNDEFINED, selectedData=Component.UNDEFINED, **kwargs):
         self._prop_names = ['animate', 'clickAnnotationData', 'style', 'figure', 'config', 'id', 'className', 'clear_on_unhover', 'relayoutData', 'clickData', 'selectedData', 'animation_options', 'hoverData']
         self._type = 'Graph'
         self._namespace = 'dash_core_components'
         self._valid_wildcard_attributes =            []
-        self.available_events = ['click', 'clickannotation', 'hover', 'selected', 'relayout', 'unhover']
-        self.available_properties = ['animate', 'clickAnnotationData', 'style', 'figure', 'config', 'id', 'className', 'clear_on_unhover', 'relayoutData', 'clickData', 'selectedData', 'animation_options', 'hoverData']
+        self.available_properties = ['id', 'clickData', 'clickAnnotationData', 'hoverData', 'clear_on_unhover', 'selectedData', 'relayoutData', 'figure', 'style', 'className', 'animate', 'animation_options', 'config']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')