+++ dash_core_components/Location.py
@@ -11,20 +11,15 @@ Use in conjunction with the `dash_core_components.Link` component to make apps w
 Keyword arguments:
 - search (string; optional): search in window.location - e.g., "?myargument=1"
 - hash (string; optional): hash in window.location - e.g., "#myhash"
-- refresh (boolean; optional): Refresh the page when the location is updated?
 - href (string; optional): href in window.location - e.g., "/my/full/pathname?myargument=1#myhash"
-- pathname (string; optional): pathname in window.location - e.g., "/my/full/pathname"
-- id (string; required)
-
-Available events: """
+- refresh (boolean; optional): Refresh the page when the location is updated?"""
     @_explicitize_args
     def __init__(self, search=Component.UNDEFINED, hash=Component.UNDEFINED, refresh=Component.UNDEFINED, href=Component.UNDEFINED, pathname=Component.UNDEFINED, id=Component.REQUIRED, **kwargs):
         self._prop_names = ['search', 'hash', 'refresh', 'href', 'pathname', 'id']
         self._type = 'Location'
         self._namespace = 'dash_core_components'
         self._valid_wildcard_attributes =            []
-        self.available_events = []
-        self.available_properties = ['search', 'hash', 'refresh', 'href', 'pathname', 'id']
+        self.available_properties = ['id', 'pathname', 'search', 'hash', 'href', 'refresh']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')