+++ dash_core_components/LogoutButton.py
@@ -31,19 +31,14 @@ Keyword arguments:
 - style (dict; optional): Style of the button
 - logout_url (string; optional): Url to submit a post logout request.
 - method (string; optional): Http method to submit the logout form.
-- className (string; optional): CSS class for the button.
-- label (string; optional): Text of the button
-- id (string; optional): Id of the button.
-
-Available events: """
+- className (string; optional): CSS class for the button."""
     @_explicitize_args
     def __init__(self, style=Component.UNDEFINED, logout_url=Component.UNDEFINED, id=Component.UNDEFINED, className=Component.UNDEFINED, label=Component.UNDEFINED, method=Component.UNDEFINED, **kwargs):
         self._prop_names = ['style', 'logout_url', 'label', 'className', 'method', 'id']
         self._type = 'LogoutButton'
         self._namespace = 'dash_core_components'
         self._valid_wildcard_attributes =            []
-        self.available_events = []
-        self.available_properties = ['style', 'logout_url', 'label', 'className', 'method', 'id']
+        self.available_properties = ['id', 'label', 'logout_url', 'style', 'method', 'className']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')