+++ dash_core_components/Store.py
@@ -19,17 +19,16 @@ Keyword arguments:
 memory: only kept in memory, reset on page refresh.
 local: window.localStorage, data is kept after the browser quit.
 session: window.sessionStorage, data is cleared once the browser quit.
-- id (string; required): The key of the storage.
-
-Available events: """
+- data (dict | list | number | string | boolean; optional): The stored data for the id.
+- clear_data (boolean; optional): Set to true to remove the data contained in `data_key`.
+- modified_timestamp (number; optional): The last time the storage was modified."""
     @_explicitize_args
     def __init__(self, modified_timestamp=Component.UNDEFINED, storage_type=Component.UNDEFINED, clear_data=Component.UNDEFINED, data=Component.UNDEFINED, id=Component.REQUIRED, **kwargs):
         self._prop_names = ['modified_timestamp', 'storage_type', 'clear_data', 'data', 'id']
         self._type = 'Store'
         self._namespace = 'dash_core_components'
         self._valid_wildcard_attributes =            []
-        self.available_events = []
-        self.available_properties = ['modified_timestamp', 'storage_type', 'clear_data', 'data', 'id']
+        self.available_properties = ['id', 'storage_type', 'data', 'clear_data', 'modified_timestamp']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')