+++ dash_core_components/Tab.py
@@ -16,21 +16,15 @@ Keyword arguments:
 - label (string; optional): The tab's label
 - className (string; optional): Appends a class to the Tab component.
 - selected_className (string; optional): Appends a class to the Tab component when it is selected.
-- selected_style (dict; optional): Overrides the default (inline) styles for the Tab component when it is selected.
-- disabled_style (dict; optional): Overrides the default (inline) styles when disabled
-- id (string; optional): The ID of this component, used to identify dash components
-in callbacks. The ID needs to be unique across all of the
-components in an app.
-
-Available events: """
+- style (dict; optional): Overrides the default (inline) styles for the Tab component.
+- selected_style (dict; optional): Overrides the default (inline) styles for the Tab component when it is selected."""
     @_explicitize_args
     def __init__(self, children=None, className=Component.UNDEFINED, style=Component.UNDEFINED, disabled_className=Component.UNDEFINED, value=Component.UNDEFINED, id=Component.UNDEFINED, disabled=Component.UNDEFINED, selected_className=Component.UNDEFINED, selected_style=Component.UNDEFINED, label=Component.UNDEFINED, disabled_style=Component.UNDEFINED, **kwargs):
         self._prop_names = ['children', 'disabled', 'style', 'disabled_className', 'value', 'label', 'className', 'selected_className', 'selected_style', 'disabled_style', 'id']
         self._type = 'Tab'
         self._namespace = 'dash_core_components'
         self._valid_wildcard_attributes =            []
-        self.available_events = []
-        self.available_properties = ['children', 'disabled', 'style', 'disabled_className', 'value', 'label', 'className', 'selected_className', 'selected_style', 'disabled_style', 'id']
+        self.available_properties = ['children', 'id', 'label', 'value', 'disabled', 'disabled_style', 'disabled_className', 'className', 'selected_className', 'style', 'selected_style']
         self.available_wildcard_properties =            []
 
         _explicit_args = kwargs.pop('_explicit_args')